Refactor webRTCConn to its own file
